+++ check_argv.py
@@ -1,0 +1,4 @@
+import sys
+
+def test_argv():
+    print("sys.argv:", sys.argv)